requirements.txt
streamlit==1.29.0
openai==1.3.8
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.1.4
Pillow==10.1.0
pytesseract==0.3.10
opencv-python==4.8.1.78
easyocr==1.7.0
pymupdf==1.23.9
openpyxl==3.1.2
langdetect==1.0.9
streamlit-chat==0.1.1
streamlit-option-menu==0.3.6
extra-streamlit-components==0.1.60
requests==2.31.0
httpx==0.25.2
